Sacados los parametros que ahora sobran de los trapecios y paralelogramos.
Movidas las imagenes a su propia carpeta texturas.

+++ tallerDeProgramacionI-SnowBros/headers/Vista/LTexture.h
@@ -29,7 +29,7 @@ class LTexture
 		void setAlpha( Uint8 alpha );
 
 		//Renders texture at given point
-		void render( SDL_Renderer* gRenderer, float x, float y, float w, float h, SDL_Rect* clip = nullptr, double angle = 0.0, SDL_Point* center = nullptr, SDL_RendererFlip flip = SDL_FLIP_NONE );
+		void render( SDL_Renderer* gRenderer, int x, int y, int w, int h, SDL_Rect* clip = nullptr, double angle = 0.0, SDL_Point* center = nullptr, SDL_RendererFlip flip = SDL_FLIP_NONE );
 
 		//The actual hardware texture
 		SDL_Texture* mTexture;
